+++ examples/cttc-error-model.cc
@@ -245,10 +245,14 @@ main (int argc, char *argv[])
   mmWaveHelper->SetSchedulerAttribute ("StartingMcsDl", UintegerValue (mcs));
   mmWaveHelper->SetSchedulerAttribute ("StartingMcsUl", UintegerValue (mcs));
 
-  // Error Model
-  mmWaveHelper->SetGnbSpectrumAttribute ("ErrorModelType", TypeIdValue (TypeId::LookupByName(errorModel)));
-  mmWaveHelper->SetUeSpectrumAttribute ("ErrorModelType", TypeIdValue (TypeId::LookupByName(errorModel)));
+  // Error Model: UE and GNB with same spectrum error model.
+  mmWaveHelper->SetUlErrorModel (errorModel);
+  mmWaveHelper->SetDlErrorModel (errorModel);
 
+  // Both DL and UL AMC will have the same model behind.
+  // Note: NOT USED in this example. MCS is fixed.
+  mmWaveHelper->SetGnbDlAmcAttribute ("AmcModel", EnumValue (NrAmc::ShannonModel));
+  mmWaveHelper->SetGnbUlAmcAttribute ("AmcModel", EnumValue (NrAmc::ShannonModel));
 
   uint32_t bwpId = 0;