+++ model/mmwave-mac-scheduler-ofdma.cc
@@ -251,8 +251,8 @@ MmWaveMacSchedulerOfdma::CreateDlDci (MmWaveMacSchedulerNs3::PointInFTPlane *spo
 {
   NS_LOG_FUNCTION (this);
 
-  uint32_t tbs = m_amc->CalculateTbSize (ueInfo->m_dlMcs,
-                                         ueInfo->m_dlRBG * GetNumRbPerRbg ());
+  uint32_t tbs = m_dlAmc->CalculateTbSize (ueInfo->m_dlMcs,
+                                           ueInfo->m_dlRBG * GetNumRbPerRbg ());
   NS_ASSERT_MSG (ueInfo->m_dlRBG % maxSym == 0, " MaxSym " << maxSym << " RBG: " << ueInfo->m_dlRBG);
   NS_ASSERT (ueInfo->m_dlRBG <= maxSym * GetBandwidthInRbg ());
   NS_ABORT_IF (maxSym > UINT8_MAX);